feat(contract-deploy): remove gateway tests and set max fee from policy
