--- packages/sdk/src/modules/deploy/index.ts
@@ -1,1 +1,0 @@
-export * from './deploy';
