+++ packages/sdk/src/modules/index.ts
@@ -1,4 +1,3 @@
 export * from './transfers';
 export * from './vault';
 export * from './auth';
-export * from './deploy';
