+++ packages/sdk/src/modules/vault/Vault.ts
@@ -222,7 +222,8 @@ export class Vault extends Predicate<[]> implements IVault {
           data: data.data.map((tx) => {
             return {
               resume: tx.resume,
-              witnesses: tx.witnesses
+              witnesses: tx.witnesses,
+              type: tx.type,
             };
           })
         };
