+++ packages/sway/test/utils/constants.ts
@@ -8,7 +8,6 @@ export const MAX_FEE = 100000000;
 export const GAS_PRICE = 1;
 
 export const CHAIN_URL = 'http://localhost:4000/v1/graphql';
-export const GATEWAY_URL = 'http://localhost:4444/v1/graphql';
 
 // transactions signed by webauthn are not posible to be signed an tested by code
 export const WEBAUTHN = {
