+++ packages/test-utils/src/utils/deploy/index.ts
@@ -1,12 +1,49 @@
-import { bn, Predicate, Provider } from 'fuels';
+import { BN, bn, BytesLike, ContractFactory, DeployContractOptions, JsonAbi, Predicate, Provider } from 'fuels';
 import { join } from 'path';
 import { readFileSync } from 'fs';
-import { BakoContractDeploy, BakoSafe } from 'bakosafe';
 import { ContractAbi__factory } from '../../types/sway';
+
+export class BakoContractDeploy extends ContractFactory {
+  readonly abi: JsonAbi;
+  readonly predicate: string;
+
+  constructor(
+    bytecode: BytesLike,
+    abi: JsonAbi,
+    provider: Provider,
+    predicate: string,
+  ) {
+    super(bytecode, abi, provider);
+    this.predicate = predicate;
+    this.abi = abi;
+  }
+
+  async deploy(deployContractOptions: DeployContractOptions = {}) {
+    if (!this.predicate) {
+      throw new Error('Predicate not set');
+    }
+
+    const { configurableConstants } = deployContractOptions;
+
+    if (configurableConstants) {
+      this.setConfigurableConstants(configurableConstants);
+    }
+
+    const { contractId, transactionRequest } = this.createTransactionRequest(
+      deployContractOptions,
+    );
+
+    return {
+      contractId,
+      transactionRequest,
+    };
+  }
+}
 
 export const createTransactionDeploy = async (
   provider: Provider,
-  vault: Predicate<any>
+  vault: Predicate<any>,
+  maxFee: BN | number,
 ) => {
   const byteCodePath = join(
     __dirname,
@@ -31,7 +68,7 @@ export const createTransactionDeploy = async (
     }
   ]);
   transactionRequest.addResources(coins);
-  transactionRequest.maxFee = bn(BakoSafe.getGasConfig('MAX_FEE'));
+  transactionRequest.maxFee = bn(maxFee);
 
   return {
     transactionRequest,
